Several styling changes, including green Antares.
Wrap all pages in a <div> which has classes set to the directory components of
the current page.  Use that to apply specific styles to pages with an "antares"
directory component.

In the course of trying to figure this out, I ended up adding the relbar on the
left for debugging, and I liked the look enough to keep it.

All of this only works if I put a hidden toc in index.rst, but that's cleaner
than the orphan declarations anyway.